Rename `pkg/ecdsa` to `pkg/tecdsa`
This package is actually about threshold-ECDSA.
Renaming it seems to be a good move since we
avoid confusion with native `crypto/ecdsa` package,
and we will achieve a better naming of package
types.
