+++ pkg/tecdsa/dkg/gen/pb/message.proto
