+++ pkg/tecdsa/dkg/marshaling.go
@@ -3,8 +3,8 @@ package dkg
 import (
 	"fmt"
 	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
-	"github.com/keep-network/keep-core/pkg/ecdsa/dkg/gen/pb"
 	"github.com/keep-network/keep-core/pkg/protocol/group"
+	"github.com/keep-network/keep-core/pkg/tecdsa/dkg/gen/pb"
 )
 
 // Marshal converts this ephemeralPublicKeyMessage to a byte array suitable for
