+++ pkg/tecdsa/dkg/result.go
